+++ allauth/account/forms.py
@@ -355,9 +355,9 @@ class SignupForm(BaseSignupForm):
                 and "password2" in self.cleaned_data:
             if self.cleaned_data["password1"] \
                     != self.cleaned_data["password2"]:
-                raise forms.ValidationError(_("You must type the same password"
-                                              " each time."))
-
+                self.add_error(
+                    'password2',
+                    _("You must type the same password each time."))
         return self.cleaned_data
 
     def save(self, request):